requirements.txt
requests
python-dotenv
streamlit
plotly
pandas
numpy
scikit-learn